+++ core/steps/__init__.py
@@ -12,6 +12,7 @@ from .condition      import ConditionStep
 from .spark_sql      import SparkSqlStep
 from .compress       import CompressStep
 from .sqoop_export   import SqoopExportStep
+from .sqoop_import   import SqoopImportStep
 from .gateway_parallel import GatewayParallelStep
 from .gateway_join   import GatewayJoinStep
 
@@ -29,6 +30,7 @@ _REGISTRY: dict[str, type[BaseStep]] = {
     "SPARK_SQL":      SparkSqlStep,
     "COMPRESS":       CompressStep,
     "SQOOP_EXPORT":   SqoopExportStep,
+    "SQOOP_IMPORT":   SqoopImportStep,
     "GATEWAY_PARALLEL": GatewayParallelStep,
     "GATEWAY_JOIN":      GatewayJoinStep,
 }