+++ ui/step_editor/__init__.py
@@ -21,6 +21,7 @@ from .condition_config_dialog import _ConditionConfigDialog
 from .spark_sql_config_dialog import _SparkSqlConfigDialog
 from .compress_config_dialog import _CompressConfigDialog
 from .sqoop_export_config_dialog import _SqoopExportConfigDialog
+from .sqoop_import_config_dialog import _SqoopImportConfigDialog
 from .gateway_parallel_config_dialog import _GatewayParallelConfigDialog
 from .gateway_join_config_dialog import _GatewayJoinConfigDialog
 
@@ -61,6 +62,7 @@ def _open_config_dialog(step_type: str, config: dict, parent,
         "SPARK_SQL":      _SparkSqlConfigDialog,
         "COMPRESS":       _CompressConfigDialog,
         "SQOOP_EXPORT":   _SqoopExportConfigDialog,
+        "SQOOP_IMPORT":   _SqoopImportConfigDialog,
         "GATEWAY_PARALLEL": _GatewayParallelConfigDialog,
         "GATEWAY_JOIN":      _GatewayJoinConfigDialog,
     }